+++ R/get_diagnoses.R
@@ -46,7 +46,7 @@
 #'
 get_diagnoses <- function(
 	codes_df,
-	file_paths = ukbrapR:::ukbrapr_paths,
+	file_paths = NULL,
 	verbose = FALSE
 )  {
 	
@@ -75,14 +75,17 @@ get_diagnoses <- function(
 	
 	# Is this one of my systems? If so, get the internal file_paths 
 	nodename <- as.character(Sys.info()['nodename'])
-	if ( nodename %in% c("SNOW","SHAPTER") )  {
+	if ( is.null(file_paths)  &  nodename %in% c("SNOW","SHAPTER") )  {
 		file_paths = ukbrapR:::snow_paths
 		if (verbose)  cli::cli_alert_info("Identified server {nodename} - using predefined paths.")
 	}
-	if ( nodename == "indy.ex.ac.uk" )  {
+	if ( is.null(file_paths)  &  nodename == "indy.ex.ac.uk" )  {
 		file_paths = ukbrapR:::indy_paths
 		if (verbose)  cli::cli_alert_info("Identified server {nodename} - using predefined paths.")
 	}
+	
+	# if file_paths not provided assume default paths
+	if (is.null(file_paths))  file_paths = ukbrapR:::ukbrapr_paths
 	
 	
 	#########################################################################################################